# cmd/server/openapi.yaml
openapi: 3.1.0
info:
  title: Echo API
  version: 1.0.0
  description: |
    The Echo feedback platform API.

    All request bodies are JSON; all responses are JSON. Authentication uses a
    Bearer JWT issued by `POST /v1/login` and carried in the `Authorization`
    header as `Bearer <access_token>`. Routes marked with a lock icon require
    this token; public routes do not.

    The dev server listens on `http://localhost:1323`.
servers:
  - url: http://localhost:1323
    description: Local development server
...
paths:
...
  /v1/invitation:
    post:
      tags: [Invitations]
      summary: Create an invitation token
      description: |
        Issues a signed invitation token that lets the bearer register as a
        member of the inviter's organization. The caller must have
        `role: "org_admin"`; any other role returns `403`. The organization is
        taken from the caller's JWT (not the body), so an admin cannot mint
        join-tokens for an organization they do not administer. The returned
        `token` is then passed as `invite_token` in `POST /v1/register` to
        create the invitee's account with `role: "user"`.
      operationId: createInvitation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
            examples:
              defaultExpiry:
                value: {}
              customExpiry:
                value:
                  expires_at: "2026-09-16T00:00:00Z"
      responses:
        "201":
          description: The signed invitation token and its metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Caller is not an org admin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notAdmin:
                  value: { message: "only org admins can create invitations" }
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            unauthorized:
              value: { message: "missing or invalid token" }
    ValidationError:
      description: |
        Malformed request body or a failed validation. The `message` carries
        the specific reason (e.g. `"name is required"`,
        `"end_date must be after start_date"`,
        `"period_id does not refer to an existing feedback period"`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalidBody:
              value: { message: "invalid request body" }
            validation:
              value: { message: "name is required" }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            internal:
              value: { message: "failed to register employee" }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A signed JWT obtained from `POST /v1/login`. Send as
        `Authorization: Bearer <access_token>`.
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable description of the error.
      required: [message]
...
    CreateInvitationRequest:
      type: object
      description: |
        Request body for `POST /v1/invitation`. The organization the invitee
        will join is taken from the inviter's JWT and cannot be set by the
        client, so an admin cannot mint join-tokens for an organization they
        do not administer.
      properties:
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: |
            ISO 8601 timestamp overriding the default 7-day lifetime. If
            omitted the token expires after 7 days. Must be in the future.
      required: []
      example:
        expires_at: "2026-09-16T00:00:00Z"
    InvitationResponse:
      type: object
      description: |
        The signed invitation token and its metadata. The
        `organization_name` is the inviter's organization (from the JWT), not
        a client-supplied value.
      properties:
        token:
          type: string
          description: The signed JWT the bearer presents at registration.
        id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
          description: The inviter employee ID.
        organization_name:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
      required: [token, id, created_by, organization_name, created_at, expires_at]
      example:
        token: eyJhbGciOiJIUzI1NiIs...
        id: 0190abcd-0000-7000-8000-000000000000
        created_by: 0190aaaa-0000-7000-8000-000000000000
        organization_name: Acme
        created_at: "2026-08-18T10:00:00Z"
        expires_at: "2026-08-25T10:00:00Z"
...